refactor: slim delegate runtime coordinator hotspots
